feat: Remove scipy as a dependency
This PR reimplements the boxcar, Hann, and Blackman window functions from scipy in pure numpy + Python, thus removing scipy as a dependency from spectre-core. A ValueError is also manually added to the get_window factory function if the window_size parameter is negative to match scipy's _len_guard helper in the original external helpers that were removed.

(Moreover, casting eltype to numpy.float32 as indicated in the type hints is enforced.)